# app/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Powerlifting API",
    "description": "API for managing athletes and records",
    "version": "1.0.0"
  },
  "paths": {
    "/token": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "operationId": "login_token_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Health Check",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/athletes": {
      "get": {
        "tags": [
          "Athletes"
        ],
        "summary": "Get Athletes",
        "operationId": "get_athletes_v1_athletes_get",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Athletes V1 Athletes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Create Athlete",
        "operationId": "create_athlete_v1_athletes_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AthleteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Athlete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/athletes/search": {
      "get": {
        "tags": [
          "Queries"
        ],
        "summary": "Search Athlete By Name",
        "operationId": "search_athlete_by_name_v1_athletes_search_get",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Name or partial name",
              "title": "Name"
            },
            "description": "Name or partial name"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Athlete"
                  },
                  "title": "Response Search Athlete By Name V1 Athletes Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/athletes/category/{weight_class}": {
      "get": {
        "tags": [
          "Queries"
        ],
        "summary": "Get By Weight Class",
        "operationId": "get_by_weight_class_v1_athletes_category__weight_class__get",
        "parameters": [
          {
            "name": "weight_class",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Weight Class"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Athlete"
                  },
                  "title": "Response Get By Weight Class V1 Athletes Category  Weight Class  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/athletes/{athlete_id}": {
      "get": {
        "tags": [
          "Athletes"
        ],
        "summary": "Get Athlete",
        "operationId": "get_athlete_v1_athletes__athlete_id__get",
        "parameters": [
          {
            "name": "athlete_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Athlete Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Athlete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete Athlete",
        "operationId": "delete_athlete_v1_athletes__athlete_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "athlete_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Athlete Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/reset": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Reset Db",
        "operationId": "reset_db_v1_reset_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "Athlete": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "age": {
            "type": "integer",
            "title": "Age"
          },
          "country": {
            "type": "string",
            "title": "Country"
          },
          "achievements": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Achievements",
            "default": []
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/Lift"
            },
            "type": "array",
            "title": "Records",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name",
          "category",
          "age",
          "country",
          "id"
        ],
        "title": "Athlete"
      },
      "AthleteCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "age": {
            "type": "integer",
            "title": "Age"
          },
          "country": {
            "type": "string",
            "title": "Country"
          },
          "achievements": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Achievements",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "name",
          "category",
          "age",
          "country"
        ],
        "title": "AthleteCreate"
      },
      "Body_login_token_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^password$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "format": "password",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "format": "password",
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_token_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Lift": {
        "properties": {
          "exercise": {
            "type": "string",
            "title": "Exercise",
            "example": "Squat"
          },
          "weight_kg": {
            "type": "number",
            "title": "Weight Kg",
            "example": 300.0
          },
          "reps": {
            "type": "integer",
            "title": "Reps",
            "example": 1
          }
        },
        "type": "object",
        "required": [
          "exercise",
          "weight_kg",
          "reps"
        ],
        "title": "Lift"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      }
    }
  }
}